hiking_app/trail-runner/requirements.txt
google-cloud-firestore==2.*
google-cloud-storage==2.*
requests==2.*